fix(koth): guard detectInitialScores against trivial division; extract HillBar
- Add inGame > 1 check in detectInitialScores to prevent the trivial
  finalCumulative % 1 === 0 fallthrough that falsely detects pre-film
  captures when per-location tick counts are uneven across hills
- Add regression test: T0 captures 3 hills with 5+5+7=17 ticks; without
  the guard the match-end capture would be dropped from hillCaptureTimestamps
- Extract HillBar component to hill-bar/hill-bar.tsx with its own tests
  subfolder; remove the WHAT comment from MIN_CAPTURE_TICKS
